initial login db fix

+++ src/main/java/com/czerwo/reworktracking/ftrot/auth/ApplicationUserRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ApplicationUserRepository extends JpaRepository<ApplicationUser, Integer> {
+public interface ApplicationUserRepository extends JpaRepository<ApplicationUser, Long> {
 
     Optional<ApplicationUser> findByUsername(String username);
